Small fix to ID

+++ tokickortostart/src/tokickortostart/Project.java
@@ -82,7 +82,7 @@ public class Project implements Comparable<Project> {
 		return this.URL;
 	}
 	
-	public double getID() {
+	public int getID() {
 		return this.ID;
 	}
